+++ CHANGELOG.md
@@ -49,6 +49,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 * [#294](https://github.com/babylonlabs-io/finality-provider/pull/294) chore: Improve fpd start
 * [#297](https://github.com/babylonlabs-io/finality-provider/pull/297) Add new command to validate pop
 * [#302](https://github.com/babylonlabs-io/finality-provider/pull/302) Update pop commands to write to a file
+* [#301](https://github.com/babylonlabs-io/finality-provider/pull/301) chore: check tx index enabled
 
 ## v0.14.3